+++ messages/de/export.php
@@ -29,6 +29,8 @@ return [
     'agenda_item'           => 'TOP',
     'prefix_short'          => 'Antragsnr.',
     'initiator'             => 'Antragsteller*in',
+    'author'                => 'Autor',
+    'date'                  => 'Datum',
     'responsibility'        => 'Zuständigkeit',
     'text'                  => 'Text',
     'tags'                  => 'Themen',