+++ messages/en/export.php
@@ -29,6 +29,8 @@ return [
     'agenda_item'           => 'Agenda',
     'prefix_short'          => 'Signature',
     'initiator'             => 'Proposer',
+    'author'                => 'Author',
+    'date'                  => 'Date',
     'responsibility'        => 'Responsibility',
     'text'                  => 'Text',
     'tags'                  => 'Tags',